Convert UI to import Tag Pairs from PGN files and merge references.
New modules in pgn-read provide ability to import from PGN files.

The UI provides a way of declaring different references in Tag Pairs to
be references to the same person or event, and correcting declarations
if errors are made.  Time controls and Modes ('over the board' and
'online' for example) are handled this way too.

The new UI is not yet connected to the existing, unchanged, performance
calculation code.

New dependency version 5.2.dev0 of solentware-base.

New dependency version 2.4.dev0 of pgn-read.

New version 2.0.0.dev0 of chesscalc.