Adjust player identification code to fit index definition changes.
A name index replaced a repr([name, ...]) index to simplify moving to
a selected point in a list sorted by name.  This broke the player
identification code which is now fixed.

The problem is moved to specifying a player in a calculation rule.

That seems the end of the disruption because a calculation rule based
on event names succeeds, as it did before.