Filter games stored for performance calculation by some PGN tag values.
Do not store game if tag pair is present and:

BlackType or WhiteType tag value is not 'human',
FEN is present and is not the starting position of a normal game,
Black or White tag value contains ':',
Result tag value is none of '1-0', '0-1', and '1/2-1/2'.

This rules out games played by computers, chess variants such as 960,
consultation games, and games without a result suitable for calculation.