Add merge players with same name option, and some ignore game reasons.
Normally identical names in different events are not seen as references
to the same person.  It may be quicker to assume the opposite and take
corrective action if needed.

Add the following reasons to not store a game's headers:

Black or White tag value is one of '-' and '?',
Termination tag value is one of 'default' and 'bye',
Any tag in Seven Tag Roster, except Round, is missing or has value '""'.

Note 'default', 'bye', and variants, as the Black or White tag value are
not seen as reasons to not store the game headers.  The same is true for
values which are clearly names of chess engines (when eyeballed).