Suppress user actions during long transactions.
Addition of LMDB to supported database engines leads to exceptions due
to incorrect transaction references when extra actions are done: such
as changing size of application window while a transaction is active.
The difference that matters is read-only transactions.

Although the other database engines are not affected they get the
treatment anyway.

More change will happen because some actions should be moved to separate
threads or processes.