Games are indexed by 'BlackType', 'Terminate', and 'WhiteType', PGN tags.
Tabs are added to display the values found: 'BlackType' and 'WhiteType'
are shown on one tab 'Player type' where expected values, from PGN
standard, are 'computer' and 'human'; some values are suggested for
'Terminate' in the PGN standard but 'default', not mentioned, is an
example of a useful value here.

The fieldname 'gamestatus' is not needed and is deleted.

The fieldnames 'gameblacktype' and 'gamewhitetype', recently added, are
replaced by 'gameplayertype'.

Version is changed, to 2.0.0.dev6, because database definition is
incompatible with earlier versions.