Cope with two database interfaces which work less well with threading.
Potentially long tranactions were put in a separate thread so the main
thread can refresh the user interface at frequent intervals.

In both cases my code tries to do things which are not allowed.

Python packages sqlite3, and tkinter (for Berkeley DB), are affected;
but apsw (SQLite3), bsddb3 (older versions of Python and Berkeley DB),
and berkeleydb, are not.